updated frontend text
I modified the text of the frontend to align better with the narrative of our submission at ETHGlobal ... AGAIN the initial part of this frontend has been produced before the hacakthon and is not the actual submission. Please see smart contracts and other novel parts to see what is made for the hackathon.

This frontend is only needed for visual presentation purposes of the functionality that has been build for ethglobal paris 2023